refactor: miscellaneous updates to typeahead Make more generic and flexible, Allow for passing scoped slot for special formatting
